pyproject.toml
[project]
name = "agentic-ai-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.1.3",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.3",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.15",
    "langchainhub>=0.1.21",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "tavily-python>=0.7.17",
]
